@@ -12,23 +12,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^5.2.5",
-    "@angular/compiler": "^5.2.5",
-    "@angular/core": "^5.2.5",
-    "@angular/forms": "^5.2.5",
-    "@angular/platform-browser": "^5.2.5",
-    "@angular/platform-browser-dynamic": "^5.2.5",
-    "@angular/router": "^5.2.5",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "bootstrap": "^4.0.0",
     "core-js": "^2.5.3",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.5.2",
     "tslib": "^1.9.0",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.9.1",
+    "rxjs-compat": "^6.0.0-rc.0"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.800.0",
     "@angular/cli": "^8.0.2",
-    "@angular/compiler-cli": "^5.2.5",
+    "@angular/compiler-cli": "^8.0.0",
     "@types/jasmine": "^2.8.6",
     "@types/node": "^9.4.6",
     "codelyzer": "^5.0.1",
